chore: relocate components/environments files to their respective new locations
