+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/attributes/components/SecondNavbar/index.tsx
